lots of changes including changing all doubles to ints and work on collision detection

+++ src/MathUtil.java
@@ -10,8 +10,8 @@ public class MathUtil {
 	 * @param angle
 	 * @return
 	 */
-	static public double normaliseAngle(double angle) {
-		return ((angle + 180.0) % 360.0 + 360.0) % 360.0 - 180.0;
+	static public int normaliseAngle(int angle) {
+		return (int) (((angle + 180.0) % 360.0 + 360.0) % 360.0 - 180.0);
 	}
 
 	/**
@@ -23,7 +23,7 @@ public class MathUtil {
 	 * @return
 	 */
 
-	public static double clamp(double value, double min, double max) {
+	public static int clamp(int value, int min, int max) {
 		return Math.max(min, Math.min(max, value));
 	}
 
@@ -36,9 +36,9 @@ public class MathUtil {
 	 *            A 3x3 matrix
 	 * @return
 	 */
-	public static double[][] multiply(double[][] p, double[][] q) {
+	public static int[][] multiply(int[][] p, int[][] q) {
 
-		double[][] m = new double[3][3];
+		int[][] m = new int[3][3];
 
 		for (int i = 0; i < 3; i++) {
 			for (int j = 0; j < 3; j++) {
@@ -61,9 +61,9 @@ public class MathUtil {
 	 *            A 3x1 vector
 	 * @return
 	 */
-	public static double[] multiply(double[][] m, double[] v) {
+	public static int[] multiply(int[][] m, int[] v) {
 
-		double[] u = new double[3];
+		int[] u = new int[3];
 
 		for (int i = 0; i < 3; i++) {
 			u[i] = 0;
@@ -80,8 +80,8 @@ public class MathUtil {
 	 * @param pos
 	 * @return
 	 */
-	public static double[][] translationMatrix(double[] v) {
-		double[][] matrix = new double[3][3];
+	public static int[][] translationMatrix(int[] v) {
+		int[][] matrix = new int[3][3];
 		matrix[0][0] = 1;
 		matrix[0][1] = 0;
 		matrix[0][2] = v[0];
@@ -101,13 +101,13 @@ public class MathUtil {
 	 *            in degrees
 	 * @return
 	 */
-	public static double[][] rotationMatrix(double angle) {
-		double[][] matrix = new double[3][3];
-		matrix[0][0] = Math.cos(Math.toRadians(angle));
-		matrix[0][1] = -Math.sin(Math.toRadians(angle));
+	public static int[][] rotationMatrix(int angle) {
+		int[][] matrix = new int[3][3];
+		matrix[0][0] = (int) Math.cos(Math.toRadians(angle));
+		matrix[0][1] = (int) -Math.sin(Math.toRadians(angle));
 		matrix[0][2] = 0;
-		matrix[1][0] = Math.sin(Math.toRadians(angle));
-		matrix[1][1] = Math.cos(Math.toRadians(angle));
+		matrix[1][0] = (int) Math.sin(Math.toRadians(angle));
+		matrix[1][1] = (int) Math.cos(Math.toRadians(angle));
 		matrix[1][2] = 0;
 		matrix[2][0] = 0;
 		matrix[2][1] = 0;
@@ -121,8 +121,8 @@ public class MathUtil {
 	 * @param scale
 	 * @return
 	 */
-	public static double[][] scaleMatrix(double scale) {
-		double[][] matrix = new double[3][3];
+	public static int[][] scaleMatrix(int scale) {
+		int[][] matrix = new int[3][3];
 		matrix[0][0] = scale;
 		matrix[0][1] = 0;
 		matrix[0][2] = 0;
